requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
requests==2.33.0
urllib3==2.6.3
beautifulsoup4
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.4.3
pandas==3.0.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.7
requests==2.33.0
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
